requirements.txt
streamlit==1.25.0
pdfplumber==0.10.2
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.0.74
Pillow==9.5.0
python-Levenshtein==0.22.0
gTTS==2.3.2
pandas==1.5.3
openpyxl==3.1.2
python-docx==0.8.11